requirements.txt
nameko==2.7.0
peewee==2.10.2
PyMySQL==0.7.11
lxml==4.1.0
pycrypto==2.6.1
fire
pytest==3.2.0
ujson
requests==2.18.4
futures
# fabric
redis==2.10.6
